requirements.txt
attrs
scipy
numpy
lxml
pygal
spacy
flask
pandas
wtforms
PyYaml
appdirs
cached_property
scikit-learn